feature: moved provider to the lib directory
I suppose it should be there.
