+++ src/main.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import RootLayout from "@/app/layout";
-import { WorkingModeProvider } from "@/providers/WorkingModeProvider";
+import { WorkingModeProvider } from "@/lib/providers/WorkingModeProvider";
 import App from "./App";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
